+++ indoxhub/__init__.py
@@ -51,7 +51,7 @@ from .exceptions import (
     ResembleBusinessPlanError,
 )
 
-__version__ = "0.2.0"
+__version__ = "0.2.1"
 __all__ = [
     "Client",
     "IndoxHub",